split bindings::common into memory and error modules
